36 - Fix cycle time (#90)
* Update date reference in cycle time calculation

 getDateInState was previously attempting to find data in `story-state` div which would not occur for a state other than the one currently active. Split the function out into separate utilities.

* Update tests

* Refactor reference to getDateInState

* Include refactored file
