+++ src/js/developmentTime/getDateInState.js
@@ -6,15 +6,9 @@ export function getDateInState(state) {
         return null;
     }
 
-    let stateDiv = document.querySelector('.story-state');
-    if (stateDiv) {
-        let stateSpan = stateDiv.querySelector('.value');
-        if (stateSpan && stateSpan.textContent !== state) {
-            return null;
-        }
-    }
-
     const parentDiv = latestUpdateElements.element.parentElement;
     const dateElement = parentDiv.querySelector('.date');
-    return dateElement ? dateElement.innerHTML : null; // Safeguard against null dateElement.
+    return dateElement ? dateElement.innerHTML : null;
 }
+
+
